% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_et_graph_bloc_communal.R
\name{data_bc_dep_equip}
\alias{data_bc_dep_equip}
\title{Le bloc communal pour les GFP}
\usage{
data_bc_dep_equip(echantillon, entite, gr_ref)
}
\arguments{
\item{echantillon}{bdd de l'échantillon sélectionné}

\item{entite}{bdd de l'entite selectionnée}

\item{gr_ref}{bdd du groupe de reference}
}
\description{
Permet d'obtenir une bdd qui va permettre de faire un graphique entre les gfp et les communes membres
}
