% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_et_graph_DRI_dep_reg.R
\name{data_DRI}
\alias{data_DRI}
\title{Etape data pour les DRI des dep et reg}
\usage{
data_DRI(echantillon, entite, gr_ref, bp_ba)
}
\arguments{
\item{echantillon}{Base de l'echnatillon GFP, com,dep,reg}

\item{entite}{Base de l'entite selectionnable type gfp ou com etc}

\item{gr_ref}{Base du gr de ref}

\item{bp_ba}{permet de choisir le input$bp_ba qui va bien}
}
\value{
une bdd
}
\description{
Permet d'obetinr la base du graphique sur les DRI pour les dep et reg
}
