% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_et_graph_evol.R
\name{data_grap_evol}
\alias{data_grap_evol}
\title{Etape data pour le graph sur les evolutions}
\usage{
data_grap_evol(echantillon, entite, gr_de_ref, bp_ba, nom_gfp_com, ma_var)
}
\arguments{
\item{echantillon}{bdd de l'echantillon}

\item{entite}{bdd de l'entite selectionne}

\item{gr_de_ref}{bdd du gr de ref}

\item{bp_ba}{la bonne valeur de l'input bp ba ou de la value}

\item{nom_gfp_com}{choix entre "Commune", "Groupement" etc...}

\item{ma_var}{selectionne la bonne variable}
}
\description{
Permet d'obetenit la base pour le graphique sur les evolutions
}
