% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_equip_graph_equip.R
\name{data_graph_equip}
\alias{data_graph_equip}
\title{Etape data pour graph equip}
\usage{
data_graph_equip(echantillon, entite, gr_ref, gfp_ou_com)
}
\arguments{
\item{echantillon}{Base de l'echnatillon GFP, com,dep,reg}

\item{entite}{Base de l'entite selectionnable type gfp ou com etc}

\item{gr_ref}{Base du gr de ref}

\item{gfp_ou_com}{Permet de travailler soit sur les gfp 1, soit sur les com 2}
}
\value{
une bdd
}
\description{
Permet d'obtenir la bonne base de donnees pour le graphique sur les depenses d'equipement
}
