% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_et_grpah_financement.R
\name{data_graph_finan}
\alias{data_graph_finan}
\title{Etape data grph financement}
\usage{
data_graph_finan(echantillon, entite, gr_de_ref, nom_gfp_com,
  annee_encours, bp_ba)
}
\arguments{
\item{echantillon}{bdd de l'echantillon}

\item{entite}{bdd de l'entite selectionnee gfp ou com ou dep ou reg}

\item{gr_de_ref}{bdd du groupe de reference de l'ofgl}

\item{nom_gfp_com}{var qui permet d'indiquer si l'entite est gfp ou commune}

\item{annee_encours}{permet de selectionne l'annee}

\item{bp_ba}{choisir le bon input de bp_ba}
}
\description{
Permet d'obtenir la bddd pour le grap de financement
}
