% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_stat_dep_equip.R
\name{data_stat_echan}
\alias{data_stat_echan}
\title{Information sur dep_equip}
\usage{
data_stat_echan(echantillon, gfp_ou_com)
}
\arguments{
\item{echantillon}{bdd de l'echantillon}

\item{gfp_ou_com}{choisir 1 pour com 2 pour gfp}
}
\description{
Permet d'obtenir un tableau pour avoir des inforamtions de dispersions sur les dep_equip
}
