% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_et_graph_detailsub.R
\name{data_sub_204}
\alias{data_sub_204}
\title{Etape data pour le details des sub}
\usage{
data_sub_204(echantillon, entite, gr_ref, annee_encours)
}
\arguments{
\item{echantillon}{Base de l'echnatillon GFP, com,dep,reg}

\item{entite}{Base de l'entite selectionnable type gfp ou com etc}

\item{gr_ref}{Base du gr de ref}

\item{annee_encours}{permet de choisir les annees}
}
\value{
une bdd
}
\description{
Permet d'obetinr la base du graphique sur les subventions
}
