% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_et_graph_sub_dep_equip.R
\name{data_sub_dep_equip}
\alias{data_sub_dep_equip}
\title{Data nuage de point}
\usage{
data_sub_dep_equip(annee_choisi, nom_dep_reg, echantillon, entite,
  gr_de_ref)
}
\arguments{
\item{annee_choisi}{permet de savoir quelle année}

\item{nom_dep_reg}{"Departement" ou "Region" etc...}

\item{echantillon}{bdd de l'echantillon}

\item{entite}{bdd de l'entite selectionnee}

\item{gr_de_ref}{bdd du groupe de ref de l'ofgl}
}
\description{
Permet d'obtenir la base pour le graph nuage de point
}
