% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_et_graph_zoom_dep_equip.R
\name{data_zoom_dep_equip}
\alias{data_zoom_dep_equip}
\title{Data pour bdd graph detail dep_equip}
\usage{
data_zoom_dep_equip(echantillon, entite, gr_de_ref, bp_ba, nom_gfp_com,
  annee_encours)
}
\arguments{
\item{echantillon}{bdd de l'echantillon}

\item{entite}{bdd de l'entite selectionee}

\item{gr_de_ref}{bdd du gr de ref}

\item{bp_ba}{choisir le bon input bp_ba}

\item{nom_gfp_com}{Commune ou groupement ou autre}

\item{annee_encours}{permet de fixer les années}
}
\description{
Permet d'obtenir la bdd qui va bien avec la fonction graph_zoom_dep_equip
}
