% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_et_graph_zoom_rec_invest.R
\name{data_zoom_rec_invest}
\alias{data_zoom_rec_invest}
\title{Data zoom sur Rec Invest}
\usage{
data_zoom_rec_invest(echantillon, entite, gr_de_ref, bp_ba, nom_gfp_com)
}
\arguments{
\item{echantillon}{bdd de l'echantillon}

\item{entite}{bdd de l'entite selectionnee}

\item{gr_de_ref}{bdd du groupe de ref de l'ofgl}

\item{bp_ba}{selectionne le bon calcul entre BP Ba ou BP + BA}

\item{nom_gfp_com}{"Commune" ou "Groupement" etc...}
}
\description{
Permet d'obtenir la base pour le graph sur les rec_invest
}
