% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_et_graph_bloc_communal.R
\name{graph_bc_dep_equip}
\alias{graph_bc_dep_equip}
\title{Graphique poids des gfp}
\usage{
graph_bc_dep_equip(data, entite, nom_gfp_com)
}
\arguments{
\item{data}{bdd pour obtenir le graphique souvent data_bc_dep_equip}

\item{entite}{bdd de l'entité permet de savoir si une entité a été isolé}

\item{nom_gfp_com}{choisir entre commune et groupement}
}
\description{
Permet de connaitre le poids des gfp dans le bloc communal
}
