% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_equip_graph_equip.R
\name{graph_dep_equip}
\alias{graph_dep_equip}
\title{Grahpique des depenses d'equipements}
\usage{
graph_dep_equip(data, entite, nom_gfp_com, gfp_ou_com, etiquette,
  gfp_ou_bc)
}
\arguments{
\item{data}{Base de donnee, souvent resultat de la fonction data_graph_equip}

\item{entite}{base de la com ou du gfp}

\item{nom_gfp_com}{Commune ou groupement}

\item{gfp_ou_com}{1 pour commune 2 pour gfp}

\item{etiquette}{permet de selectionne la base pour le total}

\item{gfp_ou_bc}{permet de modifier le titre}
}
\description{
Grahpique des depenses d'equipements
}
