% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_et_graph_DRI_dep_reg.R
\name{graph_dep_invest}
\alias{graph_dep_invest}
\title{Grahpique des depenses d'equipements}
\usage{
graph_dep_invest(data, entite, nom_gfp_com, etiquette)
}
\arguments{
\item{data}{Base de donnee, souvent resultat de la fonction data_graph_equip}

\item{entite}{base de la com ou du gfp}

\item{nom_gfp_com}{Commune ou groupement}

\item{etiquette}{permet de selectionne la base pour le total}
}
\description{
Grahpique des depenses d'equipements
}
