% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_et_graph_evol.R
\name{graph_evol}
\alias{graph_evol}
\title{Graph des evolutions des variables}
\usage{
graph_evol(data_evol, annee_dispo, bp_ba, var_select, entite)
}
\arguments{
\item{data_evol}{la bdd pour le grph souvent data_graph_evol}

\item{annee_dispo}{les annee disponible sous la forme c(2012,2013)...}

\item{bp_ba}{permet de choisir les titre en fonction d'un input}

\item{var_select}{modifie le titre du graph en fonction d'un input}

\item{entite}{bdd de l'entite selectionne}
}
\description{
Permet d'obtenir le graph des evolutions
}
