% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_et_grpah_financement.R
\name{graph_financement}
\alias{graph_financement}
\title{Graphique sur les moyens de financement}
\usage{
graph_financement(data_graph, entite, nom_gfp_com, bp_ba)
}
\arguments{
\item{data_graph}{bdd qui permet de faire le graphique data_grph_finan}

\item{entite}{bdd de l'entite selectionnee}

\item{nom_gfp_com}{var qui permet d'indiquer si l'entite est gfp ou commune}

\item{bp_ba}{permet de choisir entre bp ba}
}
\description{
Permet de faire le graphque sur les moyens de financement
}
