% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_et_graph_detailsub.R
\name{graph_sub_204_reg}
\alias{graph_sub_204_reg}
\title{Grahpique des subvention}
\usage{
graph_sub_204_reg(data, entite, nom_gfp_com, quel_annee)
}
\arguments{
\item{data}{Base de donnee, souvent resultat de la fonction data_graph_equip}

\item{entite}{base de la com ou du gfp}

\item{nom_gfp_com}{Commune ou groupement}

\item{quel_annee}{permet de choisir si 2017 ou 2012-2017}
}
\description{
Grahpique des subvention
}
