% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_et_graph_sub_dep_equip.R
\name{graph_sub_dep_equip}
\alias{graph_sub_dep_equip}
\title{Graphique pour le nuage de point}
\usage{
graph_sub_dep_equip(annee, data_graph, entite)
}
\arguments{
\item{annee}{Permet de savoir sur quelles annees on travaille}

\item{data_graph}{bdd pour le graph, doit être associé à data_sub_dep_equip}

\item{entite}{bdd de l'entite}
}
\description{
Permet d'obtenir le nuage de point entre dep_equip et subvention
}
