% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_et_graph_zoom_dep_equip.R
\name{graph_zoom_dep_equip}
\alias{graph_zoom_dep_equip}
\title{Graphique zoom dep_equip}
\usage{
graph_zoom_dep_equip(data_graph, entite, nom_gfp_com, bp_ba)
}
\arguments{
\item{data_graph}{bdd pour le graph utilisez data_zoom_dep_equip}

\item{entite}{bdd de l'entite, permet de savoir si une entite a été selectionné}

\item{nom_gfp_com}{Permet de choisir entre groupement ou autre pour le graphique}

\item{bp_ba}{selectionne le bon subtitle}
}
\description{
Permet d'obtenir le graphique qui va bien avec data_zoom_dep_equip
}
