% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_et_graph_zoom_rec_invest.R
\name{graph_zoom_rec_invest}
\alias{graph_zoom_rec_invest}
\title{Graph sur le zoom sur les rec_invest}
\usage{
graph_zoom_rec_invest(data_graph, entite, nom_gfp_com, bp_ba)
}
\arguments{
\item{data_graph}{bdd qui permet d'obtenir le graph}

\item{entite}{bdd de l'entite, permet de savoir si une entite a été selectionné}

\item{nom_gfp_com}{Permet de choisir entre groupement ou autre pour le graphique}

\item{bp_ba}{pour le subtitle}
}
\description{
Permet d'obtenir le graph sur les details des rec_invest
}
